+++ packages/stacked/lib/src/view_model_builder.dart
@@ -1,4 +1,3 @@
-import 'package:flutter/scheduler.dart';
 import 'package:flutter/widgets.dart';
 import 'package:provider/provider.dart';
 import 'base_view_models.dart';
@@ -35,6 +34,13 @@ class ViewModelBuilder<T extends ChangeNotifier> extends StatefulWidget {
 
   final _ViewModelBuilderType providerType;
 
+  /// Indicates if the onModelReady should fire every time the model is inserted into the widget tree.
+  /// Or only once during the lifecycle of the model.
+  final bool fireOnModelReadyOnce;
+
+  /// Indicates if we should run the initialise functionality for special viewmodels only once
+  final bool initialiseSpecialViewModelsOnce;
+
   /// Constructs a viewmodel provider that will not rebuild the provided widget when notifyListeners is called.
   ///
   /// Widget from [builder] will be used as a staic child and won't rebuild when notifyListeners is called
@@ -44,8 +50,12 @@ class ViewModelBuilder<T extends ChangeNotifier> extends StatefulWidget {
     this.onModelReady,
     this.disposeViewModel = true,
     this.createNewModelOnInsert = false,
+    this.fireOnModelReadyOnce = false,
+    this.initialiseSpecialViewModelsOnce = false,
+    Key key,
   })  : providerType = _ViewModelBuilderType.NonReactive,
-        staticChild = null;
+        staticChild = null,
+        super(key: key);
 
   /// Constructs a viewmodel provider that fires the [builder] function when notifyListeners is called in the viewmodel.
   const ViewModelBuilder.reactive({
@@ -55,7 +65,11 @@ class ViewModelBuilder<T extends ChangeNotifier> extends StatefulWidget {
     this.onModelReady,
     this.disposeViewModel = true,
     this.createNewModelOnInsert = false,
-  }) : providerType = _ViewModelBuilderType.Reactive;
+    this.fireOnModelReadyOnce = false,
+    this.initialiseSpecialViewModelsOnce = false,
+    Key key,
+  })  : providerType = _ViewModelBuilderType.Reactive,
+        super(key: key);
 
   @override
   _ViewModelBuilderState<T> createState() => _ViewModelBuilderState<T>();
@@ -83,32 +97,29 @@ class _ViewModelBuilderState<T extends ChangeNotifier>
       _model = widget.viewModelBuilder();
     }
 
-    _initialiseSpecialViewModels();
+    if (widget.initialiseSpecialViewModelsOnce &&
+        !(_model as BaseViewModel).initialised) {
+      _initialiseSpecialViewModels();
+      (_model as BaseViewModel)?.setInitialised(true);
+    } else if (!widget.initialiseSpecialViewModelsOnce) {
+      _initialiseSpecialViewModels();
+    }
 
     // Fire onModelReady after the model has been constructed
     if (widget.onModelReady != null) {
-      widget.onModelReady(_model);
+      if (widget.fireOnModelReadyOnce &&
+          !(_model as BaseViewModel).onModelReadyCalled) {
+        widget.onModelReady(_model);
+        (_model as BaseViewModel)?.setOnModelReadyCalled(true);
+      } else if (!widget.fireOnModelReadyOnce) {
+        widget.onModelReady(_model);
+      }
     }
   }
 
   void _initialiseSpecialViewModels() {
-    // Add any additional actions here for spcialised ViewModels
-    // TODO: Provide a closed implemenation of this functionality. Refer to the Open Closed
-    // principle in the SOLID principles
-    if (_model is FutureViewModel) {
-      (_model as FutureViewModel).runFuture();
-    }
-
-    if (_model is MultipleFutureViewModel) {
-      (_model as MultipleFutureViewModel).runFutures();
-    }
-
-    if (_model is StreamViewModel) {
-      (_model as StreamViewModel).initialise();
-    }
-
-    if (_model is MultipleStreamViewModel) {
-      (_model as MultipleStreamViewModel).initialise();
+    if (_model is Initialisable) {
+      (_model as Initialisable).initialise();
     }
   }
 
@@ -158,5 +169,5 @@ class _ViewModelBuilderState<T extends ChangeNotifier>
   }
 }
 
-/// EXPERIMENTAL: Returns the ViewModel provided above this widget in the tree 
+/// EXPERIMENTAL: Returns the ViewModel provided above this widget in the tree
 T getParentViewModel<T>(BuildContext context) => Provider.of<T>(context);