+++ packages/stacked_services/example/lib/main.dart
@@ -1,3 +1,5 @@
+import 'package:example/ui/setup_dialog_ui.dart';
+import 'package:example/ui/setup_snackbar_ui.dart';
 import 'package:flutter/material.dart';
 import 'package:stacked_services/stacked_services.dart';
 
@@ -6,6 +8,8 @@ import 'app/routes.gr.dart';
 
 void main() {
   setupLocator();
+  setupDialogUi();
+  setupSnackbarUi();
   runApp(MyApp());
 }